closes #1864
