+++ lib/riot+compiler.js
@@ -1,4 +1,4 @@
-import * as riot from './browser/index'
+import * as riot from './riot'
 import { compile, parsers } from './browser/compiler/index'
 import { extend } from './browser/util'
